pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paramiko-cloud"
version = "1.3.0"
description = "Use cloud-managed keys to sign SSH certificates"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
  { name = "Jason Rigby", email = "hello@jasonrig.by" },
]
dependencies = [
  "paramiko",
  "cryptography",
  "protobuf",
  "grpcio-tools",
]

[project.optional-dependencies]
aws = ["boto3"]
gcp = ["google-cloud-kms"]
azure = ["azure-keyvault-keys", "azure-identity"]
all = [
  "boto3",
  "google-cloud-kms",
  "azure-keyvault-keys",
  "azure-identity",
]

[project.urls]
Homepage = "https://github.com/jasonrig/paramiko-cloud/"

[tool.setuptools]
include-package-data = true
packages = [
  "paramiko_cloud",
  "paramiko_cloud.dummy",
  "paramiko_cloud.aws",
  "paramiko_cloud.azure",
  "paramiko_cloud.gcp",
  "paramiko_cloud.protobuf",
]

[dependency-groups]
dev = ["pytest", "pytest-cov", "ruff", "mypy"]
build = ["build"]
docs = [
  "sphinx",
  "paramiko",
  "cryptography",
  "boto3",
  "google-cloud-kms",
  "azure-keyvault-keys",
  "azure-identity",
  "protobuf",
  "grpcio-tools",
]

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
exclude = "(^|/)test_.*\\.py$|(^|/)paramiko_cloud/protobuf/.*_pb2.*\\.py$"
disable_error_code = ["import-untyped"]

[[tool.mypy.overrides]]
module = ["paramiko_cloud.protobuf.*"]
ignore_errors = true

[tool.ruff]
target-version = "py39"
